lesson14/dz/requirements.txt
allure-pytest
pytest
selenium
requests
pytest-xdist
